place DataBrew components in their own files. Other minor PR changes
