+++ tests/providers/amazon/aws/hooks/test_databrew.py
@@ -1,0 +1,38 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+from unittest import mock
+
+from airflow.providers.amazon.aws.hooks.databrew import GlueDataBrewHook
+
+if TYPE_CHECKING:
+    from unittest.mock import MagicMock
+
+
+class TestGlueDataBrewHook:
+    job_name = "test-databrew-job"
+    runId = "test12345"
+
+    @mock.patch.object(GlueDataBrewHook, "get_job_state")
+    def test_get_job_state(self, get_job_state_mock: MagicMock):
+        get_job_state_mock.return_value = "SUCCEEDED"
+        hook = GlueDataBrewHook()
+        result = hook.get_job_state(self.job_name, self.runId)
+        assert result == "SUCCEEDED"
